pyproject.toml
[project]
name = "pizza-project-drf"
version = "0.1.0"
description = ""
authors = [
    {name = "XeindsTwo",email = "you@example.com"}
]
requires-python = ">=3.13"
dependencies = [
    "djangorestframework (>=3.17.1,<4.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
